# Conflicts:
#	Market-Trading-System/src/main/java/DomainLayer/Market/User/IUserFacade.java
#	Market-Trading-System/src/main/java/DomainLayer/Market/User/UserController.java

+++ Market-Trading-System/src/main/java/DomainLayer/Market/Store/StoreController.java
@@ -1,14 +1,17 @@
 package DomainLayer.Market.Store;
 
 import DAL.ItemDTO;
-import DomainLayer.Market.Util.IRepository;
+import DomainLayer.Market.Util.*;
 import DomainLayer.Market.ShoppingBasket;
 import DomainLayer.Market.Purchase.IPurchaseFacade;
 import DomainLayer.Market.User.IUserFacade;
 import DomainLayer.Market.Util.IdGenerator;
 import DomainLayer.Market.Store.Item;
 
+
+import java.security.cert.CertificateException;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 
@@ -17,6 +20,19 @@ public class StoreController implements IStoreFacade{
     private IRepository<Long, Store> storesRepo;
     private IPurchaseFacade purchaseFacade;
     private IUserFacade userFacade;
+
+    private String VIEW_INVENTORY = "VIEW_INVENTORY";
+    private String ADD_ITEM = "ADD_ITEM";
+    private String UPDATE_ITEM = "UPDATE_ITEM";
+    private String DELETE_ITEM = "DELETE_ITEM";
+    private String ASSIGN_OWNER = "ASSIGN_OWNER";
+    private String ASSIGN_MANAGER = "ASSIGN_MANAGER";
+    private String VIEW_STORE_MANAGEMENT_INFO = "VIEW_STORE_MANAGEMENT_INFO";
+    private String VIEW_PURCHASE_HISTORY = "VIEW_PURCHASE_HISTORY";
+    private String CHANGE_POLICY = "CHANGE_POLICY";
+    private String CHANGE_DISCOUNT_TYPE = "CHANGE_DISCOUNT_TYPE";
+    private String REMOVE_STORE = "REMOVE_STORE";
+
 
     public StoreController(IRepository<Long, Store> storesRepo, IPurchaseFacade purchaseFacade, IUserFacade userFacade) {
         this.storesRepo = storesRepo;
@@ -29,42 +45,47 @@ public class StoreController implements IStoreFacade{
     }
 
     @Override
-    public void createStore(String founderId, String storeName, String storeDescription, IRepository<Long, Discount> discounts) {
-        //TODO: check the user is registered
+    public void createStore(String founderId, String storeName, String storeDescription, IRepository<Long, Discount> discounts) throws Exception{
+        if(userFacade.isRegister(founderId))
+            throw new Exception("User isn't registered, so can't create new store");
         long storeId = generateStoreId();
         Store newStore = new Store(storeId, founderId, storeName, storeDescription, discounts);
         storesRepo.save(newStore);
     }
 
     @Override
-    public List<String> viewInventoryByStoreOwner(String userId, long storeId) {
-        //TODO: check the user is the store owner- decide if he must be store owner
+    public HashMap<Long, Integer> viewInventoryByStoreOwner(String userId, long storeId)throws Exception {
+        if(userFacade.checkPermission(userId, storeId, VIEW_INVENTORY))
+            throw new Exception("User doesn't has permission to view the store inventory");
         Store store = storesRepo.findById(storeId);
         List<Item> inventory = store.viewInventory();
-        List<String> itemsInfo = new ArrayList<>();
-        for( Item item: inventory){
-            itemsInfo.add(item.getName());
+        HashMap<Long, Integer> itemsInfo = new HashMap<>();
+        for(Item item: inventory){
+            itemsInfo.put(item.getId(), item.getQuantity());
         }
         return itemsInfo;
     }
 
     @Override
-    public void addItemToStore(String userId, long storeId, String itemName, double itemPrice, int stockAmount, String description, List<String> categories) {
-        //TODO: check the user is the store owner
+    public void addItemToStore(String userId, long storeId, String itemName, double itemPrice, int stockAmount, String description, List<String> categories) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, ADD_ITEM))
+            throw new Exception("User doesn't has permission to add item to the store");
         Store store = storesRepo.findById(storeId);
-        store.addItem(itemName, itemPrice, stockAmount, description, categories);
+        store.addItem(IdGenerator.generateId(), itemName, itemPrice, stockAmount, description, categories);
     }
 
     @Override
-    public void updateItem(String userId, long storeId, long itemId, String newName, double newPrice, int stockAmount) {
-        //TODO: check the user is the store owner/manager with permissions
+    public void updateItem(String userId, long storeId, long itemId, String newName, double newPrice, int stockAmount) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, UPDATE_ITEM))
+            throw new Exception("User doesn't has permission to update item in the store");
         Store store = storesRepo.findById(storeId);
         store.updateItem(itemId, newName, newPrice, stockAmount);
     }
 
     @Override
-    public void deleteItem(String userId, long storeId, long itemId) {
-        //TODO: check the user is the store owner/manager with permissions
+    public void deleteItem(String userId, long storeId, long itemId) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, DELETE_ITEM))
+            throw new Exception("User doesn't has permission to delete item in the store");
         Store store = storesRepo.findById(storeId);
         store.deleteItem(itemId);
     }
@@ -81,52 +102,78 @@ public class StoreController implements IStoreFacade{
     }
 
     @Override
-    public void assignStoreOwner(String userId, long storeId, String newOwnerId){
-        //TODO: check the user is the store owner
+    public void assignStoreOwner(String userId, long storeId, String newOwnerId)throws Exception{
+        if(userFacade.checkPermission(userId, storeId, ASSIGN_OWNER))
+            throw new Exception("User doesn't has permission to assign store owner");
         Store store = storesRepo.findById(storeId);
-        userFacade.assignStoreOwner(userId, storeId);
+        userFacade.assignStoreOwner(newOwnerId, storeId);
         store.assignOwner(newOwnerId);
     }
 
     @Override
-    public void assignStoreManager(String userId, long storeId, String newManagerId){
-        //TODO: check the user is the store owner/manager with permissions
+    public void assignStoreManager(String userId, long storeId, String newManagerId, List<String> permissions)throws Exception{
+        if(userFacade.checkPermission(userId, storeId, ASSIGN_MANAGER))
+            throw new Exception("User doesn't has permission to assign store manager");
         Store store = storesRepo.findById(storeId);
-        userFacade.assignStoreManager(userId, storeId);
+        userFacade.assignStoreManager(newManagerId, storeId, permissions);
         store.assignManager(newManagerId);
     }
 
     @Override
-    public void removeStore(String userId, long storeId) {
-        //TODO: check the user is the founder
+    public void removeStore(String userId, long storeId) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, REMOVE_STORE))
+            throw new Exception("User doesn't has permission to remove store");
         storesRepo.delete(storeId);
     }
 
     @Override
-    public List<String> viewStoreManagementInfo(String userId, long storeId) {
-        //TODO: check the user is the store owner
+    public HashMap<String, List<String>> viewStoreManagementInfo(String userId, long storeId) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, VIEW_STORE_MANAGEMENT_INFO))
+            throw new Exception("User doesn't has permission to view the store management information");
         Store store = storesRepo.findById(storeId);
         List<String> managementIds = store.getManagers();
         managementIds.addAll(store.getOwners());
-        return managementIds;
+        HashMap<String, List<String>> managementInfo = new HashMap<>();
+        for(String id: managementIds){
+            managementInfo.put(id, userFacade.getUserPermission(id, storeId));
+        }
+        return managementInfo;
     }
 
     @Override
-    public void viewPurchaseHistory(String userId, long storeId) {
-        //TODO: check the user is the store owner
-        purchaseFacade.getPurchaseByStore(storeId);
+    public HashMap<Long, HashMap<Long, Integer>> viewPurchaseHistory(String userId, long storeId) throws Exception{
+        if(userFacade.checkPermission(userId, storeId, VIEW_PURCHASE_HISTORY))
+            throw new Exception("User doesn't has permission to view the store purchase history");
+        //TODO: alter according to purchase new signature
+        HashMap<Long, List<ItemDTO>> purchases = purchaseFacade.getPurchasesByStore(storeId);
+        HashMap<Long, HashMap<Long, Integer>> result = new HashMap<>();
+        for(Long purchase: purchases.keySet()){
+            result.put(purchase, new HashMap<>());
+            for(ItemDTO itemDTO: purchases.get(purchase)) {
+                result.get(purchase).put(itemDTO.getItemId(), itemDTO.getQuantity());
+            }
+        }
+        return result;
     }
 
     @Override
-    public HashMap<Long, HashMap<String, String>> getAllProductsInfoByStore(long storeId) {
-
-        return null;
+    public HashMap<Long, String> getAllProductsInfoByStore(long storeId) {
+        Store store = storesRepo.findById(storeId);
+        List<Item> items = store.viewInventory();
+        HashMap<Long, String> result = new HashMap<>();
+        for(Item item: items){
+            result.put(item.getId(), item.getName());
+        }
+        return result;
     }
 
     @Override
-    public HashMap<Long, HashMap<String, String>> getAllStoreInfo(long storeId) {
-
-        return null;
+    public HashMap<Long, String> getAllStoreInfo() {
+        HashMap<Long, String> result = new HashMap<>();
+        for(Store store: storesRepo.findAll()){
+            result.put(store.getId(), store.getName());
+        }
+        return result;
     }
 
     @Override
@@ -218,5 +265,29 @@ public class StoreController implements IStoreFacade{
             Store store = storesRepo.findById(itemDto.getStoreId());
             store.updateAmount(itemDto.getItemId(), itemDto.getQuantity());
         }
+    }
+
+    @Override
+    public void calculateBasketPrice(ShoppingBasket basket, String code) throws Exception {
+        Store store = storesRepo.findById(basket.getId());
+        store.calculateBasketPrice(basket, code);
+    }
+
+    @Override
+    public void addHiddenDiscount(double percent, Date expirationDate, List<Long> items, long storeId, String code, boolean isStoreDiscount) {
+        Store store = storesRepo.findById(storeId);
+        if(isStoreDiscount)
+            store.addDiscount(new HiddenDiscount(IdGenerator.generateId(), percent, expirationDate, storeId, code));
+        else
+            store.addDiscount(items, new HiddenDiscount(IdGenerator.generateId(), percent, expirationDate, storeId, code));
+    }
+
+    @Override
+    public void addRegularDiscount(double percent, Date expirationDate, List<Long> items, long storeId, List<Long> conditionItems, boolean isStoreDiscount) {
+        Store store = storesRepo.findById(storeId);
+        if(isStoreDiscount)
+            store.addDiscount(new RegularDiscount(IdGenerator.generateId(), percent, expirationDate, storeId, conditionItems));
+        else
+            store.addDiscount(items, new RegularDiscount(IdGenerator.generateId(), percent, expirationDate, storeId, conditionItems));
     }
 }
